:white_check_mark: fix patch of RichAsyncClient
